Error handling and reporting
The error handling and reporting is nearly done, just need to fix how the regular expression checks the email. Whenever an account entry is valid it redirects to band.html at the moment. If it needs to direct somewhere else just change the URL at the very top of the login.js